basic prompt vers

+++ bamazonSchema.sql
@@ -7,7 +7,7 @@ CREATE TABLE products(
   item_id INT NOT NULL AUTO_INCREMENT,
   product_name VARCHAR(20) NOT NULL,
   department_name VARCHAR(20) NOT NULL,
-  price DECIMAL default 0,
+  price FLOAT default 0,
   stock_quantity INT default 0,
   PRIMARY KEY (item_id)
 );
